requirements.txt
pytest==7.2.1
allure-pytest==2.9.45
selenium>=4.0.0
webdriver-manager>=4.0.0
retrying>=1.3.3
Faker>=20.1.0
requests>=2.28.0
pytest-html==3.2.0
python-dotenv==1.0.0